requirements.txt
adal==1.2.0
asn1crypto==0.24.0
azure==4.0.0
azure-applicationinsights==0.1.0
azure-batch==4.1.3
azure-common==1.1.16
azure-cosmosdb-nspkg==2.0.2
azure-cosmosdb-table==1.0.5
azure-datalake-store==0.0.37
azure-eventgrid==1.2.0
azure-graphrbac==0.40.0
azure-keyvault==1.1.0
azure-loganalytics==0.1.0
azure-mgmt==4.0.0
azure-mgmt-advisor==1.0.1
azure-mgmt-applicationinsights==0.1.1
azure-mgmt-authorization==0.50.0
azure-mgmt-batch==5.0.1
azure-mgmt-batchai==2.0.0
azure-mgmt-billing==0.2.0
azure-mgmt-cdn==3.0.0
azure-mgmt-cognitiveservices==3.0.0
azure-mgmt-commerce==1.0.1
azure-mgmt-compute==4.3.1
azure-mgmt-consumption==2.0.0
azure-mgmt-containerinstance==1.3.0
azure-mgmt-containerregistry==2.4.0
azure-mgmt-containerservice==4.2.2
azure-mgmt-cosmosdb==0.4.1
azure-mgmt-datafactory==0.6.0
azure-mgmt-datalake-analytics==0.6.0
azure-mgmt-datalake-nspkg==3.0.1
azure-mgmt-datalake-store==0.5.0
azure-mgmt-datamigration==1.0.0
azure-mgmt-devspaces==0.1.0
azure-mgmt-devtestlabs==2.2.0
azure-mgmt-dns==2.1.0
azure-mgmt-eventgrid==1.0.0
azure-mgmt-eventhub==2.2.0
azure-mgmt-hanaonazure==0.1.1
azure-mgmt-iotcentral==0.1.0
azure-mgmt-iothub==0.5.0
azure-mgmt-iothubprovisioningservices==0.2.0
azure-mgmt-keyvault==1.1.0
azure-mgmt-loganalytics==0.2.0
azure-mgmt-logic==3.0.0
azure-mgmt-machinelearningcompute==0.4.1
azure-mgmt-managementgroups==0.1.0
azure-mgmt-managementpartner==0.1.0
azure-mgmt-maps==0.1.0
azure-mgmt-marketplaceordering==0.1.0
azure-mgmt-media==1.0.0
azure-mgmt-monitor==0.5.2
azure-mgmt-msi==0.2.0
azure-mgmt-network==2.3.0
azure-mgmt-notificationhubs==2.0.0
azure-mgmt-nspkg==3.0.2
azure-mgmt-policyinsights==0.1.0
azure-mgmt-powerbiembedded==2.0.0
azure-mgmt-rdbms==1.5.0
azure-mgmt-recoveryservices==0.3.0
azure-mgmt-recoveryservicesbackup==0.3.0
azure-mgmt-redis==5.0.0
azure-mgmt-relay==0.1.0
azure-mgmt-reservations==0.2.1
azure-mgmt-resource==2.0.0
azure-mgmt-scheduler==2.0.0
azure-mgmt-search==2.0.0
azure-mgmt-servicebus==0.5.3
azure-mgmt-servicefabric==0.2.0
azure-mgmt-signalr==0.1.1
azure-mgmt-sql==0.9.1
azure-mgmt-storage==2.0.0
azure-mgmt-subscription==0.2.0
azure-mgmt-trafficmanager==0.50.0
azure-mgmt-web==0.35.0
azure-nspkg==3.0.2
azure-servicebus==0.21.1
azure-servicefabric==6.3.0.0
azure-servicemanagement-legacy==0.20.6
azure-storage-blob==1.4.0
azure-storage-common==1.4.0
azure-storage-file==1.4.0
azure-storage-queue==1.4.0
boto3==1.7.70
botocore==1.10.70
cachetools==3.0.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
construct==2.5.3
cryptography==2.4.1
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.1.1
django-bootstrap3==10.0.1
django-bootstrap4==0.0.6
django-crispy-forms==1.7.2
django-gcloud-storage==0.3.1
django-heroku==0.3.1
django-progressive-web-app==0.1.1
django-spa==0.2.0
django-storages==1.6.6
django-suit-redactor==0.0.4
django-taggit==0.23.0
django-taggit-serializer==0.1.7
djangorestframework==3.9.0
docutils==0.14
elasticsearch==6.3.0
elasticsearch-dsl==6.2.1
env==0.1.0
eyeD3==0.7.11
future==0.16.0
google-api-core==1.5.2
google-auth==1.6.1
google-cloud-core==0.28.1
google-cloud-storage==1.13.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.5
grako==3.99.9
gunicorn==19.9.0
heroku==0.1.4
idna==2.7
ipaddress==1.0.22
isodate==0.6.0
jmespath==0.9.3
Markdown==3.0.1
msrest==0.6.1
msrestazure==0.5.1
oauthlib==2.1.0
pathlib==1.0.1
pefile==2018.8.8
Pillow==5.1.0
pipenv==2018.7.1
protobuf==3.6.1
psycopg2==2.7.4
psycopg2-binary==2.7.5
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
PyJWT==1.6.4
pysolr==3.7.0
python-dateutil==2.7.3
python-decouple==3.1
python-magic==0.4.15
python-ptrace==0.9.3
python3-openid==3.1.0
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
rsa==4.0
s3transfer==0.1.13
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==2.0.0
urllib3==1.23
virtualenv==16.0.0
virtualenv-clone==0.3.0
whitenoise==3.2.2
